add: field loating message
데미지 표시용 플로팅 메시지

+++ Assets/Scripts/Ingame/Player/CastleWall.cs
@@ -19,7 +19,7 @@ namespace Ingame
             if (other.CompareTag("Enemy"))
             {
                 var enemy = other.GetComponent<Enemy>();
-                enemy.TakeDamage(int.MaxValue);
+                enemy.TakeDamage(float.MaxValue);
                 TakeHit(enemy.Damage);
             }
         }
